refactor: move checkout page logic to client component
